Move downloading logic from dataset to gcs
